refactor(gateway): split restored admission startup
