+++ src/gateway/server-startup-finish.ts
@@ -14,6 +14,7 @@ import { createLazyPromise } from "../shared/lazy-runtime.js";
 import { collectGatewayProcessMemoryUsageMb, finishGatewayRestartTrace } from "./restart-trace.js";
 import type { GatewayKernelRuntime } from "./server-kernel-request-runtime.js";
 import { GATEWAY_EVENTS } from "./server-methods-list.js";
+import { createRestoredAdmissionBeforeReady } from "./server-restored-admission-startup.js";
 import { getRequiredSharedGatewaySessionGeneration } from "./server-shared-auth-generation.js";
 import type { GatewayPostReadySidecarHandle } from "./server-startup-post-attach.js";
 import type { GatewayHttpTransport } from "./server-transport-bridge.js";
@@ -104,8 +105,6 @@ export async function finishGatewayStartup(params: {
     controlUiBasePath,
     controlUiRootLifecycle,
     sidecarStartup,
-    restoredStartup,
-    getRestoredOwnerReadiness,
     workerLiveEvents,
     earlyRuntime,
     cfgAtStart,
@@ -321,34 +320,7 @@ export async function finishGatewayStartup(params: {
         stopRegisteredPostReadySidecars,
         stopRegisteredGatewayLifetimeSidecars,
         unregisterGatewayLifetimeSidecar,
-        ...(restoredStartup
-          ? {
-              beforeReady: async () => {
-                const completed = await restoredStartup.complete({
-                  descriptor: restoredStartup.descriptor,
-                  startScheduler: async () => {
-                    const reconciliation = cronReconciliation.arm({
-                      reason: "startup",
-                      config: cfgAtStart,
-                      cronState: runtimeState.cronState,
-                    });
-                    await runtimeState.cronState.cron.start();
-                    cronStartState.handled = true;
-                    await reconciliation.complete();
-                  },
-                  getOwnerReadiness: getRestoredOwnerReadiness,
-                });
-                if (!restoredStartup.release()) {
-                  throw new Error("restored Gateway startup lost work admission");
-                }
-                startupState.restoredAdmissionReady = true;
-                log.info("restored admission opened", {
-                  readinessIdentity: completed.record.readinessIdentity,
-                  replayed: completed.replayed,
-                });
-              },
-            }
-          : {}),
+        ...createRestoredAdmissionBeforeReady({ runtime, log }),
         ...(workerPlacementRuntime
           ? {
               startWorkerEnvironmentRuntime: async () => {
